refactor: tweak `resolve_user_defined_entries` (#7727)
Move `resolve_user_defined_entries` into `ModuleLoader`
